SendSMS v1.0
Got texting functionality to work, but... not in the most intuitive way
unfortunately.

Next step is to get the phone numbers from the contacts the user
defines.

Then, add the tokens that denote the replacement form

Then, parse the strings and give it the ability to loop through and send
multiple texts